Add WebDavConfig
